src/python/requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
Django==2.0.6
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.14.1
opencv-python==3.4.4.19
pymongo==3.7.0
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyserial==3.4
python-dateutil==2.7.5
pytz==2018.5
requests==2.21.0
scipy==1.0.0
six==1.12.0
urllib3==1.24.1
